+++ source/gbc/Device.cpp
@@ -59,7 +59,7 @@ Device::Device()
      serialCycles(0), serialCallback(nullptr) {
 }
 
-void Device::tick(float dt) {
+void Device::tick(double dt) {
    uint64_t previousCycles = cpu.getCycles();
    uint64_t targetCycles = previousCycles + static_cast<uint64_t>(CPU::kClockSpeed * dt);
